requirements.txt
PySide6>=6.5.0
PyYAML>=6.0
psutil>=5.9.0
requests>=2.28.0
